Bundle version-specific codegen logic for parser generator
