+++ crates/codegen/runtime/generator/src/parser/scanner_definition.rs
@@ -5,7 +5,7 @@ use proc_macro2::TokenStream;
 use quote::{format_ident, quote};
 
 use crate::parser::grammar::{ScannerDefinitionNode, ScannerDefinitionRef};
-use crate::parser::parser_definition::VersionQualityRangeVecExtensions;
+use crate::parser::versioned::VersionedQuote;
 
 pub trait ScannerDefinitionExtensions {
     fn to_scanner_code(&self) -> TokenStream;
@@ -51,7 +51,7 @@ impl ScannerDefinitionNodeExtensions for ScannerDefinitionNode {
         match self {
             ScannerDefinitionNode::Versioned(body, version_quality_ranges) => {
                 let body = body.to_scanner_code();
-                Some(version_quality_ranges).wrap_code(body, Some(quote! { false }))
+                Some(version_quality_ranges).to_conditional_code(body, Some(quote! { false }))
             }
 
             ScannerDefinitionNode::Optional(node) => {
